Delete the external zsh-completions submodule
to allow it to be sourced independently. This also removes the burden of
having to update the submodule gitlink when there are new commits.

We could change the submodule to track a fixed branch instead (requires
git 1.8.2+, see https://stackoverflow.com/a/18799234/2654518), but just
removing the submodule, and letting the users handle it the way they
prefer, is simpler and more flexible.