Don't complete uninteresting users, and remove
    zstyle ':completion:*:*:cd:*' tag-order local-directories directory-stack path-directories
    zstyle ':completion:*:-tilde-:*' group-order 'named-directories' 'path-directories' 'expand'

as I don't see the use for them. First one is removing the completion
for named-directories, and second one is settting it only for
completions starting with tilde. Am I missing something here, or is this
reduntant, since named directories always start with tilde...?

And last, but not least, reorganize the code so it looks closer to the
prezto code, and give credits to the original contributors.