Fix glob used for checking if dat file is up-to-date
Glob used for listing completion functions in
fpath requires EXTENDED_GLOB option to be set.
If it isn't set, array of completion functions
will always be empty and dat file will always
be considered up-to-date.
I added logic for setting EXTENDED_GLOB option
and restoring it to previous state after
completion is initialized.