Revert "expand-or-complete-or-list-files when TAB is pressed"
This will interfere with other modules trying to bindkey '^I' too.
Also, we were maybe trying to be too smart and customize way beyond what
Zsh offers, which is what we generally avoid. Just set insert-tab to
false for the minimal customization that gets as close as possible to
the initially proposed solution in #16.

Fixes zimfw/fzf#7