Set COMPLETE_IN_WORD and update matcher-list
to a simpler syntax with arguably better experience, as it's gradually
more fuzzy. Only the last matcher changed. The updated matcher-list is
analogous to:

    zstyle ':completion:*' matcher-list \
         'm:{[:lower:]}={[:upper:]}' \
         'm:{[:lower:]}={[:upper:]} r:|[._-]=* r:|=*'
         'm:{[:lower:]}={[:upper:]} r:|[._-]=* r:|=* l:|=*'

compared to the previous:

    zstyle ':completion:*' matcher-list \
         'm:{[:lower:]}={[:upper:]}' \
         'm:{[:lower:]}={[:upper:]} r:|[._-]=* r:|=*'
         'm:{[:lower:]}={[:upper:]} l:|=* r:|=*'

The 'r|=*' makes better sense with COMPLETE_IN_WORD. By default, the
matcher includes a '*' at the cursor position, which would always be the
end of the input without COMPLETE_IN_WORD. We force the '*' at the end
with 'r|=*' and additionally COMPLETE_IN_WORD will include an extra '*'
in the middle of the input in case the cursor is anywhere there.

This is still not perfect, and completion is Zsh is complex enough
already. See https://www.zsh.org/mla/workers/2025/msg00125.html